requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.7
dbus-python==1.4.0
fabric @ git+https://github.com/Fabric-Development/fabric.git@27037055701557e42220a99d562439b918136bd6
idna==3.10
ijson==3.4.0
loguru==0.7.2
pillow==11.3.0
psutil==6.1.0
pycairo==1.27.0
PyGObject==3.50.0
PyGObject-stubs==2.10.0
pyjson5==1.6.8
python-networkmanager==2.2
pytomlpp==1.0.13
qrcode==8.2
requests==2.32.4
rlottie-python==1.3.6
setproctitle==1.3.4
six==1.17.0
urllib3==2.5.0
